# cli: Rework tracing output