Even for tiny bits of work (such as symbolization of a single address), any generated traces are extremely hard to parse by a human because of the verbosity that tracing-subscriber's FmtSubscriber introduces: for each trace event or span event literally all fields of all active spans are included in each emitted line.
There does not appear to be a reasonable way around this, other than writing a custom "layer" that does better. This change introduces such a layer and switches the program over to using it. Note that the reason for the default verbosity may be that it is somewhat necessary in concurrent/multi-threaded contexts. However, we do not live in such a world, so this constraint doesn't really apply to us. As a side effect of this rework, we also now emit traces to stderr, which makes it easier to separate them from the "regular" output, if desired. We also stop emitting precise span timings. For our use cases (and again, given the lack of concurrency), the emitted wall clock times are sufficient.

Before:
```
  $ cargo run -p blazecli -- symbolize process --pid 4545 0x557680951000 -vvvv
  > 2024-12-11T18:46:29.413956Z  INFO symbolize: new src=Process(4545) addrs=AbsAddr([0x557680951000])
  > 2024-12-11T18:46:29.414090Z  INFO symbolize: enter src=Process(4545) addrs=AbsAddr([0x557680951000])
  > 2024-12-11T18:46:29.414975Z  INFO symbolize:handle_entry_addr: new src=Process(4545) addrs=AbsAddr([0x557680951000]) addr=0x557680951000 entry=MapsEntry { range: 0x557680951000..0x557680a27000, offset: 0x6000, path: "/usr/bin/coreutils" }
  > 2024-12-11T18:46:29.415161Z  INFO symbolize:handle_entry_addr: enter src=Process(4545) addrs=AbsAddr([0x557680951000]) addr=0x557680951000 entry=MapsEntry { range: 0x557680951000..0x557680a27000, offset: 0x6000, path: "/usr/bin/coreutils" }
  > 2024-12-11T18:46:29.416618Z  INFO symbolize:handle_entry_addr:symbolize_with_resolver: new src=Process(4545) addrs=AbsAddr([0x557680951000]) addr=0x557680951000 entry=MapsEntry { range: 0x557680951000..0x557680a27000, offset: 0x6000, path: "/usr/bin/coreutils" } addr=0x6000 resolver=Dwarf("/proc/4545/map_files/557680951000-557680a27000")
  > 2024-12-11T18:46:29.416815Z  INFO symbolize:handle_entry_addr:symbolize_with_resolver: enter src=Process(4545) addrs=AbsAddr([0x557680951000]) addr=0x557680951000 entry=MapsEntry { range: 0x557680951000..0x557680a27000, offset: 0x6000, path: "/usr/bin/coreutils" } addr=0x6000 resolver=Dwarf("/proc/4545/map_files/557680951000-557680a27000")
  > 2024-12-11T18:46:29.417493Z  INFO symbolize:handle_entry_addr:symbolize_with_resolver: exit src=Process(4545) addrs=AbsAddr([0x557680951000]) addr=0x557680951000 entry=MapsEntry { range: 0x557680951000..0x557680a27000, offset: 0x6000, path: "/usr/bin/coreutils" } addr=0x6000 resolver=Dwarf("/proc/4545/map_files/557680951000-557680a27000")
  > 2024-12-11T18:46:29.417685Z  INFO symbolize:handle_entry_addr:symbolize_with_resolver: close time.busy=673µs time.idle=401µs src=Process(4545) addrs=AbsAddr([0x557680951000]) addr=0x557680951000 entry=MapsEntry { range: 0x557680951000..0x557680a27000, offset: 0x6000, path: "/usr/bin/coreutils" } addr=0x6000 resolver=Dwarf("/proc/4545/map_files/557680951000-557680a27000")
  > 2024-12-11T18:46:29.417945Z  INFO symbolize:handle_entry_addr: exit src=Process(4545) addrs=AbsAddr([0x557680951000]) addr=0x557680951000 entry=MapsEntry { range: 0x557680951000..0x557680a27000, offset: 0x6000, path: "/usr/bin/coreutils" }
  > 2024-12-11T18:46:29.418088Z  INFO symbolize:handle_entry_addr: close time.busy=2.78ms time.idle=337µs src=Process(4545) addrs=AbsAddr([0x557680951000]) addr=0x557680951000 entry=MapsEntry { range: 0x557680951000..0x557680a27000, offset: 0x6000, path: "/usr/bin/coreutils" }
  > 2024-12-11T18:46:29.418397Z  INFO symbolize: exit src=Process(4545) addrs=AbsAddr([0x557680951000])
  > 2024-12-11T18:46:29.418532Z  INFO symbolize: close time.busy=4.30ms time.idle=296µs src=Process(4545) addrs=AbsAddr([0x557680951000])
  > 0x00557680951000: <no-symbol>
```

After:
```
  $ cargo run -p blazecli -- symbolize process --pid 4545 0x557680951000 -vvvv
  > 2024-12-11T18:46:53.896003Z  INFO symbolize: new src=Process(4545) addrs=AbsAddr([0x557680951000])
  > 2024-12-11T18:46:53.896043Z  INFO symbolize: enter
  > 2024-12-11T18:46:53.896168Z  INFO symbolize:handle_entry_addr: new addr=0x557680951000 entry=MapsEntry { range: 0x557680951000..0x557680a27000, offset: 0x6000, path: "/usr/bin/coreutils" }
  > 2024-12-11T18:46:53.896187Z  INFO symbolize:handle_entry_addr: enter
  > 2024-12-11T18:46:53.896485Z  INFO symbolize:handle_entry_addr:symbolize_with_resolver: new addr=0x6000 resolver=Dwarf("/proc/4545/map_files/557680951000-557680a27000")
  > 2024-12-11T18:46:53.896498Z  INFO symbolize:handle_entry_addr:symbolize_with_resolver: enter
  > 2024-12-11T18:46:53.896607Z  INFO symbolize:handle_entry_addr:symbolize_with_resolver: exit
  > 2024-12-11T18:46:53.896623Z  INFO symbolize:handle_entry_addr: exit
  > 2024-12-11T18:46:53.896643Z  INFO symbolize: exit
  > 0x00557680951000: <no-symbol>
```